fix(cli): move log rotation back to DebugLogger with console logger callback
